renamed linearGap to constGap
